+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/TestEngineRegistry.java
@@ -21,13 +21,13 @@ public class TestEngineRegistry implements Iterable<TestEngine> {
 
 	private Map<String, TestEngine> testEnginesById;
 
-	public TestEngineRegistry(Set<String> testEngineIds) {
-		List<TestEngine> otherTestEngines = loadOtherTestEngines();
+	public TestEngineRegistry(Set<String> includedTestEngineIds, Set<String> excludedTestEngineIds) {
+		List<TestEngine> otherTestEngines = loadOtherTestEngines(excludedTestEngineIds);
 
 		// If there are no test engines set we don't need to filter but simply use all other test engines.
-		if (!testEngineIds.isEmpty()) {
+		if (!includedTestEngineIds.isEmpty()) {
 			otherTestEngines = otherTestEngines.stream()
-					.filter(testEngine -> testEngineIds.contains(testEngine.getId())).collect(
+					.filter(testEngine -> includedTestEngineIds.contains(testEngine.getId())).collect(
 							Collectors.toList());
 		}
 
@@ -35,11 +35,12 @@ public class TestEngineRegistry implements Iterable<TestEngine> {
 	}
 
 	/** Uses the {@link ServiceLoader} to discover all {@link TestEngine}s but the {@link ImpactedTestEngine}. */
-	private List<TestEngine> loadOtherTestEngines() {
+	private List<TestEngine> loadOtherTestEngines(Set<String> excludedTestEngineIds) {
 		List<TestEngine> testEngines = new ArrayList<>();
 
 		for (TestEngine testEngine : ServiceLoader.load(TestEngine.class, ClassLoaderUtils.getDefaultClassLoader())) {
-			if (!ImpactedTestEngine.ENGINE_ID.equals(testEngine.getId())) {
+			if (!ImpactedTestEngine.ENGINE_ID.equals(testEngine.getId()) && !excludedTestEngineIds.contains(
+					testEngine.getId())) {
 				testEngines.add(testEngine);
 			}
 		}